Add files via upload
Collar_wiring_diagram.jpg shows the pin connections used in the code.  The remote has the same NRF24L01+ connections, but no LEDs or piezo, and has a button between pin D2 and GND.

Collar_standalone.ino contains code snippets to make the collar blink and beep, but does not incorporate the NRF24L01+ or a remote.

Collar.ino and Remote.ino contain working code to make the collar blink and beep when the remote button is pressed.